clang-tidy fixes

+++ libs/unistdpp/poll.cpp
@@ -6,9 +6,10 @@ Result<int>
 poll(std::vector<pollfd>& pollfds,
      std::optional<std::chrono::milliseconds> timeout) {
 
-  int result = ::poll(pollfds.data(),
-                      pollfds.size(),
-                      timeout.has_value() ? timeout->count() : -1);
+  int result =
+    ::poll(pollfds.data(),
+           pollfds.size(),
+           timeout.has_value() ? static_cast<int>(timeout->count()) : -1);
   if (result == -1) {
     return tl::unexpected(getErrno());
   }
